Applying comments #1
